Attempt to optimize Day 1. Also reorganize project using cargo bin.
Day 1 used to always advance one character at a time even if it matched a whole word. For example, consider the line "eighteight". It would match on the first "eight" immediately, then it would advance one character and try to match "ighteight", then "ghteight", even though the word "eight" has already been matched on and extracted. It's not a logic error since there're no digit words that are subset of another digit word, but it is, I'm assuming, doing duplicate work. It might not have been since after `next()` the next call was `last()` which might have been optimized into a `next_back()`, which would mean this was a non-issue. I have also not benchmarked it so I don't know. But I did make it Unicode safe.