Day 3 stage 1 complete. Was a fucking stupid error of course.
I forgot to check number parsing at the end of a line. If the last character on a line is a digit then that number doesn't get added. The website is just fucked up and lied to me. The total sum of all numbers is way higher than the actual answer but it told me that it was lower for some fucking reason.