Optimize the line processing by reusing a Vec.
Instead of allocating and deallocation a new collection for each line of input,
we instead just reuse a vec through the entire program.
However, since we're dropping each line after it's processed,
in order to push substrings of the input string into the vec,
we need to use unsafe to cast the str borrow into an arbitrary lifetime.